+++ src/Cache/CacheKeyAndTagsGenerator.php
@@ -2,7 +2,6 @@
 
 namespace Nuwave\Lighthouse\Cache;
 
-use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Contracts\Auth\Authenticatable;
 
 class CacheKeyAndTagsGenerator implements CacheKeyAndTags
@@ -11,19 +10,19 @@ class CacheKeyAndTagsGenerator implements CacheKeyAndTags
     public const SEPARATOR = ':';
 
     /**
-     * @param  int|string|null $id
-     * @param  array<string, mixed> $args
+     * @param  int|string|null  $id
+     * @param  array<string, mixed>  $args
+     * @param  array<int, string> $path
      */
     public function key(
         ?Authenticatable $user,
         bool $isPrivate,
+        string $parentName,
         $id,
+        string $fieldName,
         array $args,
-        ResolveInfo $resolveInfo
+        array $path
     ): string {
-        $parentName = $resolveInfo->parentType->name;
-        $fieldName = $resolveInfo->fieldName;
-
         $parts = [self::PREFIX];
 
         if ($isPrivate && null !== $user) {